refactor: use imports with `js` file extensions
for better compatibility with TypeScript < 5.0
