+++ src/bundle/index.ts
@@ -1,6 +1,6 @@
 /* eslint-disable simple-import-sort/exports */
-export * as pdf from "swissqrbill:pdf:index";
-export * as svg from "swissqrbill:svg:index";
+export * as pdf from "swissqrbill:pdf";
+export * as svg from "swissqrbill:svg";
 export * as types from "swissqrbill:types";
 export * as utils from "swissqrbill:utils";
 export * as errors from "swissqrbill:errors";
