+++ src/pdf/index.ts
@@ -1,3 +1,3 @@
-export * from "./swissqrbill";
-export * from "./swissqrcode";
-export * from "./table";
+export * from "./swissqrbill.js";
+export * from "./swissqrcode.js";
+export * from "./table.js";
