+++ src/shared/cleaner.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import { cleanData } from "swissqrbill:shared:cleaner";
+import { cleanData } from "swissqrbill:shared:cleaner.js";
 
 import type { Data } from "swissqrbill:types";
