+++ src/shared/translations.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, test } from "vitest";
 
-import { minimalRequired } from "swissqrbill:tests:data/valid-data";
-import { pdf } from "swissqrbill:tests:utils/pdf";
-import { svg } from "swissqrbill:tests:utils/svg";
+import { minimalRequired } from "swissqrbill:tests:data/valid-data.js";
+import { pdf } from "swissqrbill:tests:utils/pdf.js";
+import { svg } from "swissqrbill:tests:utils/svg.js";
 
 
 describe("translations", async () => {
