+++ src/svg/index.ts
@@ -1,2 +1,2 @@
-export * from "./swissqrbill";
-export * from "./swissqrcode";
+export * from "./swissqrbill.js";
+export * from "./swissqrcode.js";
