+++ tests/utils/buffer.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import { splitBuffer } from "swissqrbill:tests:utils/buffer";
+import { splitBuffer } from "swissqrbill:tests:utils/buffer.js";
 
 
 describe("buffer", () => {
