+++ tests/utils/svg.ts
@@ -3,7 +3,7 @@ import { dirname, join } from "node:path";
 
 import SVG from "svg-engine";
 
-import { SwissQRBill } from "swissqrbill:svg:index";
+import { SwissQRBill } from "swissqrbill:svg";
 
 import type { Data, SVGOptions } from "swissqrbill:types";
